+++ packages/cli/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.11.0 - 2025-02-06
+
 ### Added
 
 - `getMetadata` in descriptors